pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference"
authors = ["Aleх Bark <thrtth@yandex.ru>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]
classifiers = [
  "Environment :: Console",
  "Programming Language :: Python :: 3.10"
]
repository = "https://github.com/thrtth/python-project-50"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.2"
pytest-cov = "^4.1.0"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
